+++ CHANGELOG.md
@@ -5,10 +5,11 @@
 - [Feature] Scroll on Infrared Screen
 - [Feature] Check Self Update App in Options (only for github)
 - [Feature] Fap Catalog save sort
-- [Feature] Check Self Update App in Optiond (only for github)
+- [Feature] Add metrics for faphub
 - [FIX] Use by default dark theme in Wear OS
 - [FIX] Use default splashscreen in Wear OS
 - [FIX] Fix compose layout in Wear OS
+- [FIX] Fix wearos font issue
 - [FIX] New mehanism emulate in Wear OS
 
 # 1.6.4